requirements.txt
discord.py
spotipy